requirements.txt
discord.py==2.3.2
openai==1.12.0
python-dotenv==1.0.1
aiohttp==3.9.3
async-timeout==4.0.3
google-generativeai==0.3.2
pydantic==2.6.1
pydantic-ai